+++ cli/build.rs
@@ -8,6 +8,7 @@ use deno_core::Extension;
 use deno_core::ExtensionFileSource;
 use deno_core::ExtensionFileSourceCode;
 use deno_runtime::deno_cache::SqliteBackedCache;
+use deno_runtime::deno_fs::StdFs;
 use deno_runtime::deno_kv::sqlite::SqliteDbHandler;
 use deno_runtime::permissions::PermissionsContainer;
 use deno_runtime::*;
@@ -361,8 +362,7 @@ fn create_cli_snapshot(snapshot_path: PathBuf) {
     deno_napi::deno_napi::init_ops::<PermissionsContainer>(),
     deno_http::deno_http::init_ops(),
     deno_io::deno_io::init_ops(Default::default()),
-    deno_fs::deno_fs::init_ops::<PermissionsContainer>(false),
-    deno_flash::deno_flash::init_ops::<PermissionsContainer>(false), // No --unstable
+    deno_fs::deno_fs::init_ops::<_, PermissionsContainer>(false, StdFs),
     deno_node::deno_node::init_ops::<deno_runtime::RuntimeNodeEnv>(None),
     cli::init_ops_and_esm(), // NOTE: This needs to be init_ops_and_esm!
   ];