+++ cli/tests/integration/watcher_tests.rs
@@ -1122,30 +1122,19 @@ fn test_watch_unload_handler_error_on_drop() {
   check_alive_then_kill(child);
 }
 
+#[cfg(unix)]
 #[test]
-fn run_watch_blob_urls_reset() {
-  let _g = util::http_server();
+fn test_watch_sigint() {
+  use nix::sys::signal;
+  use nix::sys::signal::Signal;
+  use nix::unistd::Pid;
+
   let t = TempDir::new();
   let file_to_watch = t.path().join("file_to_watch.js");
-  let file_content = r#"
-    const prevUrl = localStorage.getItem("url");
-    if (prevUrl == null) {
-      console.log("first run, storing blob url");
-      const url = URL.createObjectURL(
-        new Blob(["export {}"], { type: "application/javascript" }),
-      );
-      await import(url); // this shouldn't insert into the fs module cache
-      localStorage.setItem("url", url);
-    } else {
-      await import(prevUrl)
-        .then(() => console.log("importing old blob url incorrectly works"))
-        .catch(() => console.log("importing old blob url correctly failed"));
-    }
-    "#;
-  write(&file_to_watch, file_content).unwrap();
+  write(&file_to_watch, r#"Deno.test("foo", () => {});"#).unwrap();
   let mut child = util::deno_cmd()
     .current_dir(util::testdata_path())
-    .arg("run")
+    .arg("test")
     .arg("--watch")
     .arg(&file_to_watch)
     .env("NO_COLOR", "1")
@@ -1154,12 +1143,12 @@ fn run_watch_blob_urls_reset() {
     .spawn()
     .unwrap();
   let (mut stdout_lines, mut stderr_lines) = child_lines(&mut child);
-  wait_contains("first run, storing blob url", &mut stdout_lines);
-  wait_contains("finished", &mut stderr_lines);
-  write(&file_to_watch, file_content).unwrap();
-  wait_contains("importing old blob url correctly failed", &mut stdout_lines);
-  wait_contains("finished", &mut stderr_lines);
-  check_alive_then_kill(child);
+  wait_contains("Test started", &mut stderr_lines);
+  wait_contains("ok | 1 passed | 0 failed", &mut stdout_lines);
+  wait_contains("Test finished", &mut stderr_lines);
+  signal::kill(Pid::from_raw(child.id() as i32), Signal::SIGINT).unwrap();
+  let exit_status = child.wait().unwrap();
+  assert_eq!(exit_status.code(), Some(130));
 }
 
 // Regression test for https://github.com/denoland/deno/issues/15465.